+++ Lesson-II/homework/src/components/Button/Button.jsx
@@ -1,5 +1,6 @@
 import React from 'react';
 
+// eslint-disable-next-line no-unused-vars
 export function Button({ props, handleClick }) {
   return <button onClick={handleClick}>click</button>;
 }
