+++ Lesson-II/homework/src/components/Input/Input.jsx
@@ -1,0 +1,16 @@
+import React, { useRef } from 'react';
+export function Input({ props, handleChange }) {
+  const message = useRef(null);
+  // eslint-disable-next-line no-unused-vars
+  function setValue() {
+    props.changeValue(message.current.value, props.element);
+  }
+  return (
+    <input
+      data-testid="inputName"
+      ref={message}
+      type="text"
+      onChange={handleChange}
+    />
+  );
+}
